Join proxy handling into tycho-core
