+++ tycho-bundles/org.eclipse.tycho.p2.resolver.impl/META-INF/MANIFEST.MF
@@ -61,7 +61,6 @@ Import-Package: org.apache.commons.io;version="2.8.0",
 Service-Component: OSGI-INF/generator.xml,
  OSGI-INF/metadataserializable.xml,
  OSGI-INF/org.eclipse.tycho.p2.manager.ReactorRepositoryManagerImpl.xml,
- OSGI-INF/proxyfacade.xml,
  OSGI-INF/remoteAgentManager.xml,
  OSGI-INF/resolver.xml,
  OSGI-INF/targetDefinitionResolverService.xml
