--- tycho-bundles/org.eclipse.tycho.p2.resolver.impl/OSGI-INF/proxyfacade.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.eclipse.tycho.p2.ProxyServiceFacade">
-   <implementation class="org.eclipse.tycho.p2.impl.proxy.ProxyServiceFacadeImpl"/>
-   <service>
-      <provide interface="org.eclipse.tycho.core.shared.ProxyServiceFacade"/>
-   </service>
-   <reference interface="org.eclipse.core.net.proxy.IProxyService" 
-      bind="setProxyServer" 
-      unbind="unsetProxyServer"/>
-</scr:component>
